Move to Diagnostics namespace
