+++ src/Elasticsearch.Net/Connection/HttpWebRequestConnection.cs
@@ -14,6 +14,7 @@ using System.Net.Security;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
+using Elasticsearch.Net.Diagnostics;
 
 namespace Elasticsearch.Net
 {
