+++ src/Elasticsearch.Net/Responses/HttpDetails/ApiCallDetails.cs
@@ -7,6 +7,7 @@ using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Net.NetworkInformation;
 using System.Text;
+using Elasticsearch.Net.Diagnostics;
 
 namespace Elasticsearch.Net
 {
