+++ src/Elasticsearch.Net/Responses/HttpDetails/IApiCallDetails.cs
@@ -7,6 +7,7 @@ using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Diagnostics;
 using System.Net.NetworkInformation;
+using Elasticsearch.Net.Diagnostics;
 
 namespace Elasticsearch.Net
 {
