+++ tests/Tests/ClientConcepts/Troubleshooting/DebugInformation.doc.cs
@@ -12,6 +12,7 @@ using System.Threading.Tasks;
 using Elastic.Elasticsearch.Xunit.Sdk;
 using Elastic.Elasticsearch.Xunit.XunitPlumbing;
 using Elasticsearch.Net;
+using Elasticsearch.Net.Diagnostics;
 using FluentAssertions;
 using Nest;
 using Tests.Core.Client;
